pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multimodal-contradiction-project"
version = "0.1.0"
description = "COCO-derived multimodal contradiction benchmark and baselines."
readme = "IMPLEMENTATION_PLAN.md"
requires-python = ">=3.10"
dependencies = [
  "accelerate>=1.1.0",
  "matplotlib>=3.8.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "Pillow>=10.0.0",
  "PyYAML>=6.0.0",
  "scikit-learn>=1.4.0",
  "seaborn>=0.13.0",
  "tensorboard>=2.16.0",
  "torch>=2.2.0",
  "torchvision>=0.17.0",
  "tqdm>=4.66.0",
  "transformers>=4.51.3",
]

[project.optional-dependencies]
qwen = ["bitsandbytes>=0.43.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["vl_contradiction*"]

[tool.ruff]
line-length = 100
target-version = "py310"
